[ebe] Done create CRUD for product admin

+++ database/migrations/2024_09_27_103818_create_products_table.php
@@ -14,13 +14,12 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price', 10, 2);
+            $table->integer('price');
             $table->text('description');
             $table->integer('stock');
             $table->string('image');
             $table->timestamps();
         });
-
     }
 
     /**
